Computed Tomography/requirements.txt
numpy~=1.19.1
Pillow~=8.3.1